# faster listing of oriented triangles in `src/sage/combinat/cluster_algebra_quiver/mutation_type.py`