We change the algorithm used in `_triangles`. The new method ensures that a triple `(x, y, z)` is considered only once.  The idea is to do:
- consider the vertices by increasing label
- Let `x` be the current vertex and let `nx` be the neighbors of `x` with a larger label
- Consider the  pairs `(y, z)` of vertices in `nx`. We know that we have an edge between `x` and each of these vertices but we don't know which ones.
- Then when either `(y,z)` or `(z, y)` is and edge, we check if the triangle is well oriented or not.
- We repeat for the vertex after `x` in the ordering. Vertex `x` will never be considered again.

The time complexity of the previous algorithm was in $O(nm)$ and the new one is in $O(n\Delta^2)$, where $\Delta$ is the maximum (undirected) degree.

Let `old_triangles` be the previous method and `new_triangles` be the new method. We get:
```sage
sage: def foo(triangles):
....:     return set((frozenset(T), b) for T, b in triangles)
sage:
sage: %time A = old_triangles(D)
CPU times: user 187 µs, sys: 38 µs, total: 225 µs
Wall time: 229 µs
sage: %time B = new_triangles(D)
CPU times: user 115 µs, sys: 16 µs, total: 131 µs
Wall time: 134 µs
sage: len(A) == len(B) and foo(A) == foo(B)
True
sage:
sage: D = graphs.CompleteGraph(10).random_orientation()
sage: %time A = old_triangles(D)
CPU times: user 1.34 ms, sys: 1 µs, total: 1.34 ms
Wall time: 1.34 ms
sage: %time B = new_triangles(D)
CPU times: user 329 µs, sys: 18 µs, total: 347 µs
Wall time: 349 µs
sage: len(A) == len(B) and foo(A) == foo(B)
True
sage:
sage: D = graphs.CompleteGraph(50).random_orientation()
sage: %time A = old_triangles(D)
CPU times: user 7.33 s, sys: 20.3 ms, total: 7.35 s
Wall time: 7.35 s
sage: %time B = new_triangles(D)
CPU times: user 66 ms, sys: 3.03 ms, total: 69.1 ms
Wall time: 67.6 ms
sage: len(A) == len(B) and foo(A) == foo(B)
True
sage:
sage: D = graphs.CompleteGraph(60).random_orientation()
sage: %time A = old_triangles(D)
CPU times: user 25.4 s, sys: 58.1 ms, total: 25.4 s
Wall time: 25.4 s
sage: %time B = new_triangles(D)
CPU times: user 44.3 ms, sys: 1.68 ms, total: 45.9 ms
Wall time: 45 ms
sage: len(A) == len(B) and foo(A) == foo(B)
True
sage: len(A)
34220
```


### :memo: Checklist

<!-- Put an `x` in all the boxes that apply. -->

- [ ] The title is concise and informative.
- [ ] The description explains in detail what this PR is about.
- [ ] I have linked a relevant issue or discussion.
- [ ] I have created tests covering the changes.
- [ ] I have updated the documentation and checked the documentation preview.

### :hourglass: Dependencies

<!-- List all open PRs that this PR logically depends on. For example, -->
<!-- - #12345: short description why this is a dependency -->
<!-- - #34567: ... -->